% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{axis_mark_}
\alias{axis_mark_}
\alias{axis_mark_x}
\alias{axis_mark_y}
\title{Configure the Axis of the Last Mark}
\usage{
axis_mark_(chart = NULL, channel, ...)

axis_mark_x(chart = NULL, ...)

axis_mark_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Axis options such as \code{position}, \code{title}, \code{grid}, etc., or
\code{FALSE} to hide.}
}
\value{
The modified \code{g2} object.
}
\description{
Set axis options for a positional channel on the most recently added mark.
This is the mark-level counterpart to \code{\link[=axis_]{axis_()}}, useful for dual-axis
charts where individual marks need independent axes. Set to \code{FALSE} to
hide the axis on just that mark.
}
\examples{
df = data.frame(x = 1:5, a = c(1, 4, 2, 5, 3), b = c(100, 200, 150, 300, 250))
g2(df, x = 'x') |>
  mark_interval(encode = list(y = 'a')) |>
  mark_line(encode = list(y = 'b')) |>
  scale_mark_y(independent = TRUE) |>
  axis_mark_y(position = 'right', grid = FALSE)
}
