% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{axis_of}
\alias{axis_of}
\title{Configure an Axis}
\usage{
axis_of(chart, channel, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Axis options such as \code{title}, \code{labelFormatter}, \code{tickCount},
\code{grid}, etc., or \code{FALSE} to hide.}
}
\value{
The modified \code{g2} object.
}
\description{
Customise the axis for a positional channel (\code{'x'} or \code{'y'}). Set to
\code{FALSE} to hide the axis.
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_point() |>
  axis_of('x', title = 'Miles per Gallon') |>
  axis_of('y', title = 'Horsepower')
}
