% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{axis_x}
\alias{axis_x}
\title{Configure the X Axis}
\usage{
axis_x(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Axis options such as \code{title}, \code{labelFormatter}, \code{tickCount},
\code{grid}, etc., or \code{FALSE} to hide.}
}
\description{
Configure the X Axis
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  axis_x(title = 'Miles per Gallon')
}
