% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_helix}
\alias{coord_helix}
\title{Helix Coordinate System}
\usage{
coord_helix(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional options such as \code{innerRadius}, \code{outerRadius},
\code{startAngle}, \code{endAngle}, or \code{transform}.}
}
\description{
Shortcut for \code{coord_(chart, 'helix', ...)}.
}
