% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_parallel}
\alias{coord_parallel}
\title{Parallel Coordinate System}
\usage{
coord_parallel(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional options such as \code{innerRadius}, \code{outerRadius},
\code{startAngle}, \code{endAngle}, or \code{transform}.}
}
\description{
Shortcut for \code{coord_(chart, 'parallel', ...)}. Used with \code{position}
encoding for parallel coordinate plots.
}
