% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_radar}
\alias{coord_radar}
\title{Radar Coordinate System}
\usage{
coord_radar(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional options such as \code{innerRadius}, \code{outerRadius},
\code{startAngle}, \code{endAngle}, or \code{transform}.}
}
\description{
Shortcut for \code{coord_(chart, 'radar', ...)}. Used with \code{position}
encoding for radar (spider) charts.
}
