% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_radial}
\alias{coord_radial}
\title{Radial Coordinate System}
\usage{
coord_radial(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional options such as \code{innerRadius}, \code{outerRadius},
\code{startAngle}, \code{endAngle}, or \code{transform}.}
}
\description{
Shortcut for \code{coord_(chart, 'radial', ...)}. Suitable for radial bar
charts.
}
\examples{
g2(data.frame(x = c('A', 'B', 'C'), y = c(3, 7, 2)),
  x = 'x', y = 'y', color = 'x') |>
  mark_interval() |> coord_radial()
}
