% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_theta}
\alias{coord_theta}
\title{Theta Coordinate System}
\usage{
coord_theta(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional options such as \code{innerRadius}, \code{outerRadius},
\code{startAngle}, \code{endAngle}, or \code{transform}.}
}
\description{
Shortcut for \code{coord_(chart, 'theta', ...)}. Used for pie and donut
charts.
}
\examples{
g2(data.frame(x = c('A', 'B', 'C'), y = c(3, 7, 2)), y ~ x, color = ~ x) |>
  transform_('stackY') |>
  coord_theta(innerRadius = 0.5)
}
