% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{encode}
\alias{encode}
\title{Set Aesthetic Mappings}
\usage{
encode(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Named aesthetic mappings, e.g., \verb{x = var1, y = var2}.}
}
\value{
The modified \code{g2} object.
}
\description{
Map data variables to visual channels (x, y, color, size, shape, etc.).
Arguments are captured as unevaluated expressions and stored in the chart
object's \code{aesthetics} field.
}
\examples{
g2(mtcars) |> encode(x = mpg, y = hp) |> mark_point()
}
