% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet_circle}
\alias{facet_circle}
\title{Facet in a Circular Layout}
\usage{
facet_circle(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Facet encoding and options. Pass \code{position = 'var'} to specify
the faceting variable.}
}
\value{
The modified \code{g2} object.
}
\description{
Facet in a Circular Layout
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length') |>
  mark_point() |>
  facet_circle(position = 'Species')
}
