% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{g2_col_cdn}
\alias{g2_col_cdn}
\title{CDN URL for the G2 Column-Major Data Helper}
\usage{
g2_col_cdn()
}
\value{
A character string.
}
\description{
CDN URL for the G2 Column-Major Data Helper
}
\keyword{internal}
