% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{header}
\alias{header}
\title{Set the Chart Header}
\usage{
header(chart = NULL, text, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{text}{Title text string.}

\item{...}{Additional title options such as \code{subtitle}, \code{align}, \code{style}.}
}
\value{
The modified \code{g2} object.
}
\description{
Set the Chart Header
}
\examples{
g2(mtcars, hp ~ mpg) |>
  header('Motor Trend Cars', subtitle = 'mpg vs hp')
}
