% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{interact}
\alias{interact}
\title{Add an Interaction}
\usage{
interact(chart, type, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{type}{Interaction type (e.g., \code{'tooltip'}, \code{'elementHighlight'}).}

\item{...}{Additional interaction options.}
}
\value{
The modified \code{g2} object.
}
\description{
Add an Interaction
}
\examples{
g2(mtcars, x = mpg, y = hp) |>
  mark_point() |>
  interact('tooltip')
}
