% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{knit_print.g2}
\alias{knit_print.g2}
\title{Custom Printing in Knitr}
\usage{
knit_print.g2(x, ...)
}
\arguments{
\item{x}{A \code{g2} object.}

\item{...}{Ignored.}
}
\value{
A \code{knit_asis} character vector.
}
\description{
Custom Printing in Knitr
}
