% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{labels_of}
\alias{labels_of}
\title{Add Labels to the Last Mark}
\usage{
labels_of(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Label options such as \code{text} (channel name), \code{position},
\code{formatter}, \code{style}.}
}
\value{
The modified \code{g2} object.
}
\description{
Append a label configuration to the most recently added mark. Can be called
multiple times to add several label layers.
}
\examples{
df = data.frame(x = c('A', 'B', 'C'), y = c(3, 7, 2))
g2(df, x = 'x', y = 'y') |>
  mark_interval() |>
  labels_of(text = 'y', position = 'inside')
}
