% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{legend_color}
\alias{legend_color}
\title{Configure the Color Legend}
\usage{
legend_color(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Legend options such as \code{position} (\code{'top'}, \code{'bottom'}, \code{'left'},
\code{'right'}), \code{layout}, \code{title}, etc.}
}
\description{
Configure the Color Legend
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length', color = 'Species') |>
  legend_color(position = 'right')
}
