% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{legend_of}
\alias{legend_of}
\title{Configure a Legend}
\usage{
legend_of(chart, channel, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Visual channel name.}

\item{...}{Legend options such as \code{position} (\code{'top'}, \code{'bottom'}, \code{'left'},
\code{'right'}), \code{layout}, \code{title}, etc.}
}
\value{
The modified \code{g2} object.
}
\description{
Customise the legend for a visual channel (\code{'color'}, \code{'size'}, \code{'shape'},
\code{'opacity'}). Set to \code{FALSE} to hide.
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length', color = 'Species') |>
  mark_point() |>
  legend_of('color', position = 'right')
}
