% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{legend_opacity}
\alias{legend_opacity}
\title{Configure the Opacity Legend}
\usage{
legend_opacity(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Legend options such as \code{position} (\code{'top'}, \code{'bottom'}, \code{'left'},
\code{'right'}), \code{layout}, \code{title}, etc.}
}
\description{
Configure the Opacity Legend
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp', opacity = 'wt') |>
  legend_opacity(position = 'bottom')
}
