% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{legend_shape}
\alias{legend_shape}
\title{Configure the Shape Legend}
\usage{
legend_shape(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Legend options such as \code{position} (\code{'top'}, \code{'bottom'}, \code{'left'},
\code{'right'}), \code{layout}, \code{title}, etc.}
}
\description{
Configure the Shape Legend
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length', shape = 'Species') |>
  legend_shape(position = 'bottom')
}
