% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{legend_size}
\alias{legend_size}
\title{Configure the Size Legend}
\usage{
legend_size(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Legend options such as \code{position} (\code{'top'}, \code{'bottom'}, \code{'left'},
\code{'right'}), \code{layout}, \code{title}, etc.}
}
\description{
Configure the Size Legend
}
\examples{
g2(mtcars, hp ~ mpg, size = ~ wt) |>
  legend_size(position = 'bottom')
}
