% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{mark_area}
\alias{mark_area}
\title{Add an Area Mark}
\usage{
mark_area(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional mark-level options passed to G2.}
}
\description{
Add an Area Mark
}
\examples{
g2(data.frame(x = 1:10, y = cumsum(rnorm(10))), x = x, y = y) |> mark_area()
}
