% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_beeswarm}
\alias{mark_beeswarm}
\title{Add a Beeswarm Mark}
\usage{
mark_beeswarm(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Displays individual data points using force simulation to avoid overlapping,
creating a beeswarm layout. Particularly useful for visualizing distributions
within categories.
}
\examples{
g2(iris, Sepal.Width ~ Species) |> mark_beeswarm()
}
