% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_box}
\alias{mark_box}
\title{Add a Box Mark}
\usage{
mark_box(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws pre-computed box elements (for custom box plots).
}
