% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_density}
\alias{mark_density}
\title{Add a Density Mark}
\usage{
mark_density(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
A composite mark for univariate kernel density estimation visualization.
The data should contain a numeric column for density estimation and
optionally categorical columns for grouping. A \code{kde} (Kernel Density
Estimation) data transform is required to compute the density curve from
the raw data.
}
