% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_force_graph}
\alias{mark_force_graph}
\title{Add a Force Graph Mark}
\usage{
mark_force_graph(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws a force-directed graph layout.
}
