% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_gauge}
\alias{mark_gauge}
\title{Add a Gauge Mark}
\usage{
mark_gauge(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Gauge Mark
}
\examples{
g2() |>
  mark_gauge(data = list(list(value = 0.75)),
    encode = list(y = 'value', color = 'value'),
    style = list(textContent = 'Score'))
}
