% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_image}
\alias{mark_image}
\title{Add an Image Mark}
\usage{
mark_image(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Places images at data coordinates. Requires an \code{src} encoding for image URLs.
}
\examples{
df = data.frame(x = 1:2, y = 1:2)
g2(df, y ~ x) |>
  mark_image(style = list(
    src = 'https://gw.alipayobjects.com/mdn/rms_dfc253/afts/img/A*SZGfRaFPkIoAAAAAAAAAAAAAARQnAQ'
  ))
}
