% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{mark_interval}
\alias{mark_interval}
\title{Add an Interval Mark (Bar/Column Chart)}
\usage{
mark_interval(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional mark-level options passed to G2.}
}
\description{
Add an Interval Mark (Bar/Column Chart)
}
\examples{
g2(data.frame(x = c('A', 'B', 'C'), y = c(3, 7, 2)), x = x, y = y) |>
  mark_interval()
}
