% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{mark_line}
\alias{mark_line}
\title{Add a Line Mark}
\usage{
mark_line(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional mark-level options passed to G2.}
}
\description{
Add a Line Mark
}
\examples{
g2(data.frame(x = 1:10, y = rnorm(10)), x = x, y = y) |> mark_line()
}
