% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_liquid}
\alias{mark_liquid}
\title{Add a Liquid Mark}
\usage{
mark_liquid(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws a liquid fill gauge.
}
\examples{
g2() |>
  mark_liquid(data = list(list(value = 0.3)),
    encode = list(y = 'value'),
    style = list(textContent = '30\%'))
}
