% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_pack}
\alias{mark_pack}
\title{Add a Pack (Circle Packing) Mark}
\usage{
mark_pack(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Pack (Circle Packing) Mark
}
\examples{
tree_data = list(
  name = 'root', children = list(
    list(name = 'A', value = 10),
    list(name = 'B', value = 20),
    list(name = 'C', value = 15)
  )
)
g2() |>
  mark_pack(
    data = list(value = tree_data),
    encode = list(value = 'value', color = 'name')
  )
}
