% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_partition}
\alias{mark_partition}
\title{Add a Partition (Sunburst) Mark}
\usage{
mark_partition(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
A composite mark for hierarchical partition (sunburst/icicle) visualization.
The data should be a nested tree structure wrapped in a list (i.e., the
\code{value} in \code{data = list(type = 'inline', value = list(tree))} should be a
list containing the root node). Use \code{coordinate = list(type = 'theta')} for
a sunburst layout, or the default Cartesian coordinates for an icicle chart.
}
