% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{mark_point}
\alias{mark_point}
\title{Add a Point Mark (Scatter Plot)}
\usage{
mark_point(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional mark-level options passed to G2.}
}
\description{
Add a Point Mark (Scatter Plot)
}
\examples{
g2(mtcars, x = mpg, y = hp) |> mark_point()
}
