% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_polygon}
\alias{mark_polygon}
\title{Add a Polygon Mark}
\usage{
mark_polygon(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Polygon Mark
}
\examples{
df = data.frame(x = c(0, 1, 0.5), y = c(0, 0, 1))
g2(df, y ~ x) |> mark_polygon()
}
