% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_range_x}
\alias{mark_range_x}
\title{Add a Horizontal Range (rangeX)}
\usage{
mark_range_x(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Horizontal Range (rangeX)
}
\examples{
g2(mtcars, hp ~ mpg) |>
  mark_point() |>
  mark_range_x(data = list(list(x = c(15, 25))),
    style = list(fill = 'steelblue', fillOpacity = 0.15))
}
