% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_shape}
\alias{mark_shape}
\title{Add a Shape Mark}
\usage{
mark_shape(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
A custom mark whose rendering is controlled by a JavaScript render function.
}
