% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_text}
\alias{mark_text}
\title{Add a Text Mark}
\usage{
mark_text(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Text Mark
}
\examples{
df = data.frame(x = c('A', 'B', 'C'), y = c(3, 7, 2))
g2(df, x = 'x', y = 'y') |>
  mark_interval() |>
  mark_text(encode = list(text = 'y'))
}
