% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_tree}
\alias{mark_tree}
\title{Add a Tree Mark}
\usage{
mark_tree(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws a tree layout.
}
