% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_vector}
\alias{mark_vector}
\title{Add a Vector Mark}
\usage{
mark_vector(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws arrows or vectors. Useful for wind or flow field visualizations.
}
