% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{render_shiny}
\alias{render_shiny}
\title{Render a Chart in Shiny}
\usage{
render_shiny(chart, session, output_id)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{session}{The Shiny session object.}

\item{output_id}{The container element ID on the client side.}
}
\description{
Build the chart configuration and send it to the client via
\code{session$sendCustomMessage()}.
}
