% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderG2}
\alias{renderG2}
\title{Render a G2 Chart in Shiny}
\usage{
renderG2(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a \code{g2} object.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is already quoted.}
}
\value{
A render function for use with \code{\link[=g2Output]{g2Output()}}.
}
\description{
Create a reactive G2 chart for use with \code{\link[=g2Output]{g2Output()}}.
Assign it to an output slot: \code{output$ID = renderG2({...})}.
}
