% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scale_mark_}
\alias{scale_mark_}
\alias{scale_mark_x}
\alias{scale_mark_y}
\title{Configure the Scale of the Last Mark}
\usage{
scale_mark_(chart = NULL, channel, ...)

scale_mark_x(chart = NULL, ...)

scale_mark_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Scale options such as \code{independent}, \code{nice}, \code{type}, \code{domain}.}
}
\value{
The modified \code{g2} object.
}
\description{
Set scale options for a positional channel on the most recently added mark.
This is the mark-level counterpart to \code{\link[=scale_]{scale_()}}, useful for dual-axis
charts where individual marks need independent scales.
}
\examples{
df = data.frame(x = 1:5, a = c(1, 4, 2, 5, 3), b = c(100, 200, 150, 300, 250))
g2(df, x = 'x') |>
  mark_interval(encode = list(y = 'a')) |>
  mark_line(encode = list(y = 'b')) |>
  scale_mark_y(independent = TRUE) |>
  axis_mark_y(position = 'right', grid = FALSE)
}
