% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_opacity}
\alias{scale_opacity}
\title{Configure the Opacity Scale}
\usage{
scale_opacity(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Scale options passed to G2 (e.g., \code{type = 'log'}, \code{nice = TRUE},
\code{domain}, \code{range}, \code{zero = TRUE}).}
}
\description{
Configure the Opacity Scale
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp', opacity = 'wt') |>
  mark_point() |>
  scale_opacity(range = c(0.2, 1))
}
