% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_shape}
\alias{scale_shape}
\title{Configure the Shape Scale}
\usage{
scale_shape(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Scale options passed to G2 (e.g., \code{type = 'log'}, \code{nice = TRUE},
\code{domain}, \code{range}, \code{zero = TRUE}).}
}
\description{
Configure the Shape Scale
}
\examples{
g2(iris, Sepal.Length ~ Sepal.Width, shape = ~ Species) |>
  scale_shape(range = c('circle', 'square', 'triangle'))
}
