% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_size}
\alias{scale_size}
\title{Configure the Size Scale}
\usage{
scale_size(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Scale options passed to G2 (e.g., \code{type = 'log'}, \code{nice = TRUE},
\code{domain}, \code{range}, \code{zero = TRUE}).}
}
\description{
Configure the Size Scale
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp', size = 'wt') |>
  mark_point() |>
  scale_size(range = c(2, 10))
}
