% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scroll_}
\alias{scroll_}
\alias{scroll_x}
\alias{scroll_y}
\title{Add a Scrollbar}
\usage{
scroll_(chart = NULL, channel, ...)

scroll_x(chart = NULL, ...)

scroll_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Scrollbar options.}
}
\value{
The modified \code{g2} object.
}
\description{
Add a scrollbar to a positional channel for zooming/panning.
}
\examples{
df = data.frame(x = 1:100, y = cumsum(rnorm(100)))
p = g2(df, y ~ x) |> mark_line()
p |> scroll_x() |> scroll_y()
}
