% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scrollbar_}
\alias{scrollbar_}
\title{Add a Scrollbar}
\usage{
scrollbar_(chart = NULL, channel, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Scrollbar options.}
}
\value{
The modified \code{g2} object.
}
\description{
Add a Scrollbar
}
\examples{
df = data.frame(x = 1:100, y = cumsum(rnorm(100)))
g2(df, y ~ x) |>
  mark_line() |>
  scrollbar_('x')
}
