% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scrollbar_of}
\alias{scrollbar_of}
\title{Add a Scrollbar}
\usage{
scrollbar_of(chart, channel, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Scrollbar options.}
}
\value{
The modified \code{g2} object.
}
\description{
Add a Scrollbar
}
\examples{
df = data.frame(x = 1:100, y = cumsum(rnorm(100)))
g2(df, x = 'x', y = 'y') |>
  mark_line() |>
  scrollbar_of('x')
}
