% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scrollbar_x}
\alias{scrollbar_x}
\title{Add an X Scrollbar}
\usage{
scrollbar_x(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Scrollbar options.}
}
\description{
Add an X Scrollbar
}
\examples{
df = data.frame(x = 1:100, y = cumsum(rnorm(100)))
g2(df, y ~ x) |>
  mark_line() |>
  scrollbar_x()
}
