% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scrollbar_y}
\alias{scrollbar_y}
\title{Add a Y Scrollbar}
\usage{
scrollbar_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Scrollbar options.}
}
\description{
Add a Y Scrollbar
}
\examples{
df = data.frame(x = 1:100, y = cumsum(rnorm(100)))
g2(df, x = 'x', y = 'y') |>
  mark_line() |>
  scrollbar_y()
}
