% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{slider_of}
\alias{slider_of}
\title{Add a Slider}
\usage{
slider_of(chart, channel, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Slider options.}
}
\value{
The modified \code{g2} object.
}
\description{
Add a range slider to a positional channel for zooming/panning.
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_point() |>
  slider_of('x')
}
