% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{style_mark}
\alias{style_mark}
\title{Set Style on the Last Mark}
\usage{
style_mark(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Style options such as \code{fill}, \code{stroke}, \code{lineWidth},
\code{fillOpacity}, \code{strokeOpacity}.}
}
\value{
The modified \code{g2} object.
}
\description{
Set Style on the Last Mark
}
\examples{
g2(mtcars, hp ~ mpg) |>
  style_mark(fill = 'steelblue', stroke = 'white', lineWidth = 1)
}
