% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_academy}
\alias{theme_academy}
\title{Academy Theme}
\usage{
theme_academy(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional theme options merged with the type.}
}
\description{
Shortcut for \code{theme_(chart, 'academy', ...)}.
}
\examples{
g2(mtcars, hp ~ mpg) |> theme_academy()
}
