% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_classic}
\alias{theme_classic}
\alias{theme_classic_dark}
\alias{theme_light}
\alias{theme_dark}
\alias{theme_academy}
\title{Set the Chart Theme}
\usage{
theme_classic(chart = NULL, ...)

theme_classic_dark(chart = NULL, ...)

theme_light(chart = NULL, ...)

theme_dark(chart = NULL, ...)

theme_academy(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional theme options.}
}
\value{
The modified \code{g2} object.
}
\description{
Shortcut functions to apply a built-in G2 theme. Available themes:
\code{theme_classic()} (default), \code{theme_classic_dark()}, \code{theme_light()},
\code{theme_dark()}, \code{theme_academy()}.
}
\examples{
p = g2(mtcars, hp ~ mpg)

# Classic (default) theme
p |> theme_classic()

# Classic dark theme
p |> theme_classic_dark()

# Light theme
p |> theme_light()

# Dark theme
p |> theme_dark()

# Academy theme
p |> theme_academy()
}
