% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_light}
\alias{theme_light}
\title{Light Theme}
\usage{
theme_light(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional theme options merged with the type.}
}
\description{
Shortcut for \code{theme_(chart, 'light', ...)}.
}
\examples{
g2(mtcars, hp ~ mpg) |> theme_light()
}
