% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{title_}
\alias{title_}
\title{Set the Chart Title}
\usage{
title_(chart = NULL, text, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{text}{Title text string.}

\item{...}{Additional title options such as \code{subtitle}, \code{align}, \code{style}.}
}
\value{
The modified \code{g2} object.
}
\description{
Set the Chart Title
}
\examples{
g2(mtcars, hp ~ mpg) |>
  title_('Motor Trend Cars', subtitle = 'mpg vs hp')
}
