% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{tooltip_of}
\alias{tooltip_of}
\title{Configure the Tooltip}
\usage{
tooltip_of(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Tooltip options such as \code{shared}, \code{crosshairs}, \code{marker}, or
\code{FALSE} to disable.}
}
\value{
The modified \code{g2} object.
}
\description{
Set chart-level tooltip options. Mark-level tooltips can be passed via \code{...}
in \verb{mark_*()} functions.
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_point() |>
  tooltip_of(crosshairs = TRUE)
}
