% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{axis_y}
\alias{axis_y}
\title{Configure the Y Axis}
\usage{
axis_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Axis options such as \code{title}, \code{labelFormatter}, \code{tickCount},
\code{grid}, \code{position}, etc., or \code{FALSE} to hide.}
}
\description{
Configure the Y Axis
}
\examples{
g2(mtcars, hp ~ mpg) |>
  axis_y(title = 'Horsepower')
}
