% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\name{encode}
\alias{encode}
\title{Set Aesthetic Mappings}
\usage{
encode(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Named mappings as character strings, e.g.,
\verb{x = 'col1', color = 'col2'}.}
}
\value{
The modified \code{g2} object.
}
\description{
Map data columns to visual channels (x, y, color, size, shape, etc.).
Column names are specified as character strings.
}
\examples{
g2(mtcars) |> encode(x = 'mpg', y = 'hp') |> mark_point()
}
