% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet_circle}
\alias{facet_circle}
\title{Facet in a Circular Layout}
\usage{
facet_circle(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Facet encoding and options. Pass \code{position = ~var} to specify
the faceting variable. Character strings are also accepted.}
}
\value{
The modified \code{g2} object.
}
\description{
Facet in a Circular Layout
}
\examples{
g2(iris, Sepal.Length ~ Sepal.Width) |>
  facet_circle(position = ~ Species)
}
