% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet_rect}
\alias{facet_rect}
\title{Facet by a Rectangular Grid}
\usage{
facet_rect(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Facet encoding and options. Pass \code{x = ~var} and/or \code{y = ~var}
to specify the faceting variable(s). Character strings are also accepted.}
}
\value{
The modified \code{g2} object.
}
\description{
Split the chart into a grid of panels based on one or two variables,
similar to ggplot2's \code{facet_grid()} / \code{facet_wrap()}. Use the \code{x} and/or
\code{y} arguments to specify faceting variables.
}
\examples{
g2(iris, Sepal.Length ~ Sepal.Width) |>
  facet_rect(x = ~ Species)
}
