% Generated manually
\name{gglite-package}
\alias{gglite-package}
\alias{gglite}
\docType{package}
\title{gglite: Lightweight helpers for ggplot2}
\description{
The gglite package streamlines exploratory plotting by providing a themed
\code{ggplot2} initialiser and helper wrappers for common geoms. Functions are
organised following the categories outlined in Posit's Data Visualisation guide.
}
\details{
\tabular{ll}{
Package: \tab gglite\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
License: \tab MIT\cr
}
}
\author{
Seoyeon Choi (First author)\cr
Guebin Choi (Corresponding author) \email{guebin@jbnu.ac.kr}\cr
\address{
Department of Statistics (Institute of Applied Statistics), Jeonbuk National University\\
567 Baekje-daero, Deokjin-gu, Jeonju-si, Jeollabuk-do, 54896 Republic of Korea
}
}
\seealso{\code{gglite()}}
