% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglite.R
\docType{package}
\name{gglite-package}
\alias{gglite}
\alias{gglite-package}
\title{gglite: Lightweight Data Visualization via the Grammar of Graphics}
\description{
An R interface to the AntV G2 JavaScript visualization library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yihui/gglite}
  \item Report bugs at \url{https://github.com/yihui/gglite/issues}
}

}
\author{
\strong{Maintainer}: Yihui Xie \email{xie@yihui.name} (\href{https://orcid.org/0000-0003-0645-5666}{ORCID})

}
\keyword{internal}
