% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{interact}
\alias{interact}
\title{Add an Interaction}
\usage{
interact(chart = NULL, type, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{type}{Interaction type string.}

\item{...}{Additional interaction options.}
}
\value{
The modified \code{g2} object.
}
\description{
Enable an interaction behaviour on the chart. G2 interaction types include:
\code{'tooltip'}, \code{'elementHighlight'}, \code{'elementHighlightByX'},
\code{'elementHighlightByColor'}, \code{'elementSelect'}, \code{'elementSelectByX'},
\code{'elementSelectByColor'}, \code{'elementHoverScale'}, \code{'fisheye'},
\code{'chartIndex'}, \code{'legendFilter'}, \code{'legendHighlight'},
\code{'brushHighlight'}, \code{'brushXHighlight'}, \code{'brushYHighlight'},
\code{'brushFilter'}, \code{'brushXFilter'}, \code{'brushYFilter'},
\code{'sliderFilter'}, \code{'poptip'}, \code{'drillDown'}.
}
\examples{
# Tooltip on scatter plot
g2(mtcars, hp ~ mpg) |>
  interact('tooltip')

# Highlight elements on hover
g2(mtcars, hp ~ mpg, color = ~ cyl) |>
  interact('elementHighlight')

# Brush to highlight a region
g2(mtcars, hp ~ mpg) |>
  interact('brushHighlight')

# Legend filter
g2(iris, Sepal.Length ~ Sepal.Width, color = ~ Species) |>
  interact('legendFilter')
}
