% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_area}
\alias{mark_area}
\title{Add an Area Mark}
\usage{
mark_area(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add an Area Mark
}
\examples{
g2(data.frame(x = 1:5, y = c(3, 1, 4, 1, 5)), y ~ x) |>
  mark_area()
}
