% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_chord}
\alias{mark_chord}
\title{Add a Chord Mark}
\usage{
mark_chord(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws a chord diagram. Data should have \code{source}, \code{target}, and \code{value}
columns.
}
\examples{
df = data.frame(
  source = c('A', 'A', 'B'), target = c('B', 'C', 'C'),
  value = c(5, 3, 2)
)
g2(df) |>
  mark_chord(
    encode = list(source = 'source', target = 'target', value = 'value')
  )
}
