% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_density}
\alias{mark_density}
\title{Add a Density Mark}
\usage{
mark_density(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
A composite mark for kernel density estimation visualization.
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length') |>
  mark_density(
    transform = list(list(type = 'kde')),
    style = list(fill = 'steelblue', fillOpacity = 0.5)
  )
}
