% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_edge}
\alias{mark_edge}
\title{Add an Edge Mark}
\usage{
mark_edge(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Used in graph visualizations together with \code{\link[=mark_node]{mark_node()}}.
}
