% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_heatmap}
\alias{mark_heatmap}
\title{Add a Heatmap Mark}
\usage{
mark_heatmap(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
A composite mark for rendering heatmaps from point data.
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length', color = 'Petal.Length') |>
  mark_heatmap()
}
