% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_interval}
\alias{mark_interval}
\title{Add an Interval Mark (Bar / Column Chart)}
\usage{
mark_interval(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add an Interval Mark (Bar / Column Chart)
}
\examples{
# Bar chart
g2(data.frame(x = c('A', 'B', 'C'), y = c(3, 7, 2)), x = 'x', y = 'y') |>
  mark_interval()

# Stacked bar chart (using transform)
df = data.frame(
  x = rep(c('A', 'B'), each = 2), y = c(3, 2, 5, 4),
  color = rep(c('a', 'b'), 2)
)
g2(df, x = 'x', y = 'y', color = 'color') |>
  mark_interval() |> transform_('stackY')
}
