% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_line}
\alias{mark_line}
\title{Add a Line Mark}
\usage{
mark_line(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Line Mark
}
\examples{
g2(data.frame(x = 1:5, y = c(3, 1, 4, 1, 5)), y ~ x) |>
  mark_line()
}
