% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_line_y}
\alias{mark_line_y}
\title{Add a Horizontal Reference Line (lineY)}
\usage{
mark_line_y(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Horizontal Reference Line (lineY)
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_point() |>
  mark_line_y(data = list(list(y = 150)),
    style = list(stroke = 'red', lineDash = c(4, 4)))
}
