% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_node}
\alias{mark_node}
\title{Add a Node Mark}
\usage{
mark_node(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Used in graph visualizations together with \code{\link[=mark_edge]{mark_edge()}}.
}
