% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_partition}
\alias{mark_partition}
\title{Add a Partition (Sunburst) Mark}
\usage{
mark_partition(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Partition (Sunburst) Mark
}
\examples{
tree_data = list(
  name = 'root', children = list(
    list(name = 'A', value = 10, children = list(
      list(name = 'A1', value = 5), list(name = 'A2', value = 5)
    )),
    list(name = 'B', value = 20)
  )
)
g2() |>
  mark_partition(
    data = list(value = tree_data),
    encode = list(value = 'value'),
    coordinate = list(type = 'theta', innerRadius = 0.3)
  )
}
