% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_point}
\alias{mark_point}
\title{Add a Point Mark (Scatter Plot)}
\usage{
mark_point(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Point Mark (Scatter Plot)
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp', color = 'cyl') |> mark_point()
}
