% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_range}
\alias{mark_range}
\title{Add a Range Mark}
\usage{
mark_range(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Range Mark
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_point() |>
  mark_range(
    data = list(list(x = c(15, 25), y = c(100, 200))),
    style = list(fill = 'steelblue', fillOpacity = 0.15)
  )
}
