% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_range_y}
\alias{mark_range_y}
\title{Add a Vertical Range (rangeY)}
\usage{
mark_range_y(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Add a Vertical Range (rangeY)
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_point() |>
  mark_range_y(data = list(list(y = c(100, 200))),
    style = list(fill = 'orange', fillOpacity = 0.15))
}
