% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_rect}
\alias{mark_rect}
\title{Add a Rect Mark}
\usage{
mark_rect(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Mark-level options passed to G2, such as \code{data}, \code{encode},
\code{transform}, \code{style}, \code{animate}, \code{labels}, \code{tooltip}, \code{axis}, \code{legend}.}
}
\description{
Draws rectangles. Commonly used with a \code{bin} transform for 2-D histograms.
}
\examples{
# 2-D histogram using bin transform
g2(mtcars, x = 'mpg', y = 'hp') |>
  mark_rect(
    transform = list(list(type = 'bin', thresholdsX = 10, thresholdsY = 10))
  )
}
