% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{preview}
\alias{preview}
\title{Preview a Chart in the Viewer or Browser}
\usage{
preview(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional arguments passed to \code{\link[=chart_html]{chart_html()}}.}
}
\value{
The path to the temporary HTML file (invisibly).
}
\description{
Preview a Chart in the Viewer or Browser
}
