% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{print.g2}
\alias{print.g2}
\title{Preview a Chart in the Viewer or Browser}
\usage{
\method{print}{g2}(x, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=chart_html]{chart_html()}}.}

\item{chart}{A \code{g2} object.}
}
\value{
The chart object (invisibly).
}
\description{
Preview a Chart in the Viewer or Browser
}
