% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_shiny}
\alias{render_shiny}
\title{Render a Chart in Shiny}
\usage{
render_shiny(chart, session, output_id)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{session}{The Shiny session object.}

\item{output_id}{The container element ID on the client side.}
}
\description{
Render a Chart in Shiny
}
