% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_color}
\alias{scale_color}
\title{Configure the Color Scale}
\usage{
scale_color(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Scale options passed to G2 (e.g., \code{type = 'log'}, \code{nice = TRUE},
\code{domain}, \code{range}, \code{zero = TRUE}).}
}
\description{
Configure the Color Scale
}
\examples{
g2(iris, x = 'Sepal.Width', y = 'Sepal.Length', color = 'Species') |>
  scale_color(palette = 'category10')
}
