% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{scrollbar_}
\alias{scrollbar_}
\alias{scrollbar_x}
\alias{scrollbar_y}
\title{Add a Scrollbar}
\usage{
scrollbar_(chart = NULL, channel, ...)

scrollbar_x(chart = NULL, ...)

scrollbar_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{channel}{Positional channel: \code{'x'} or \code{'y'}.}

\item{...}{Scrollbar options.}
}
\value{
The modified \code{g2} object.
}
\description{
Add a scrollbar to a positional channel for zooming/panning.
}
\details{
\code{scrollbar_x()}: Shortcut for \code{scrollbar_(chart, 'x', ...)}.

\code{scrollbar_y()}: Shortcut for \code{scrollbar_(chart, 'y', ...)}.
}
\examples{
df = data.frame(x = 1:100, y = cumsum(rnorm(100)))
p = g2(df, y ~ x) |> mark_line()
p |> scrollbar_('x')

# Scrollbar shortcuts
p |> scrollbar_x()
p |> scrollbar_y()
}
