% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{slider_x}
\alias{slider_x}
\title{Add an X Slider}
\usage{
slider_x(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Slider options.}
}
\description{
Add an X Slider
}
\examples{
g2(mtcars, hp ~ mpg) |>
  slider_x()
}
