% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{slider_y}
\alias{slider_y}
\title{Add a Y Slider}
\usage{
slider_y(chart = NULL, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Slider options.}
}
\description{
Add a Y Slider
}
\examples{
g2(mtcars, hp ~ mpg) |>
  slider_y()
}
