% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_classic}
\alias{theme_classic}
\title{Classic Theme}
\usage{
theme_classic(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional theme options merged with the type.}
}
\description{
Shortcut for \code{theme_(chart, 'classic', ...)}. This is the default theme.
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |> mark_point() |> theme_classic()
}
