% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_classicDark}
\alias{theme_classicDark}
\title{Classic Dark Theme}
\usage{
theme_classicDark(chart, ...)
}
\arguments{
\item{chart}{A \code{g2} object.}

\item{...}{Additional theme options merged with the type.}
}
\description{
Shortcut for \code{theme_(chart, 'classicDark', ...)}.
}
\examples{
g2(mtcars, x = 'mpg', y = 'hp') |> mark_point() |> theme_classicDark()
}
